feature: simulate scene until a specific object emits a signal
Currently, the runner only allows to simulate until the scene object itself emits a signal. Very often however we have objects in the scene that we try to simulate and the scene itself is just a test harness. With this new function the scene runner can now simulate the scene until any object that is given to the function will emit a specific signal.