GD-380: Fixes the swallowing of input events when using `GdUnitCmdTool`
# Why
When testing a scene via the `GdUnitCmdTool` input events were swallowed by the internally used RichTextLabelExt node.
The RTL is added to the main scene and thus swallows the input events and are not passed to the scene under test.

# What
We don't need the `RichTextLabelExt` in the tool this was only designed to use in the report inspector to print colored backgrounds.
Replaced it be the Godot provided ``RichTextLabel` implementation