GD-390: Use Godot logfile path to scan for test runtime errors
# Why
Tests never ends when a user changed the path of the Godot log file to another path/name.
The `GodotGdErrorMonitor` uses a hardcoded logfile path and did not check if this path exists and ran into an
an endless loop when trying to parse the logs.

# What
Now uses the logfile path configured by Godot and added a check if the logfile exists